mock blueprint compact
